Change file extensions from .jsx to .js
