+++ src/App.js
@@ -1,8 +1,10 @@
 import React from 'react';
-import "tachyons";
+import 'tachyons';
 import './App.css';
 import { CTA, Navbar } from './components';
-import { About, FAQS, Footer, Header, HIW, Services } from './containers';
+import {
+  About, FAQS, Footer, Header, HIW, Services,
+} from './containers';
 
 const App = () => (
   <div className="App">
