+++ src/components/faq/FAQ.js
@@ -1,0 +1,16 @@
+import React from 'react';
+import './FAQ.css';
+
+const FAQ = () => (
+  <div className="w-100 w-40-l w-40-m shadow-5 pointer grow pa3 mb5">
+    <h3 className="mb3">How to subscribe for a meal</h3>
+    <p className="mb3">
+      Lorem Ipsum is simply dummy text of the printing and typesetting
+      industry. Lorem Ipsum has been the industry's standard dummy text ever
+      since the 1500s,
+    </p>
+    <button className="bg-transparent gold-text bw0 dim">READ MORE</button>
+  </div>
+);
+
+export default FAQ;
