+++ src/components/service/Service.js
@@ -1,0 +1,10 @@
+import React from 'react';
+import './Service.css';
+
+const Service = () => (
+  <div>
+    Service
+  </div>
+);
+
+export default Service;
