--- src/components/service/Service.jsx
@@ -1,12 +1,0 @@
-import React from 'react';
-import './Service.css';
-
-const Service = () => {
-  return (
-    <div>
-      Service
-    </div>
-  )
-}
-
-export default Service
