+++ src/slides.js
@@ -1,28 +1,29 @@
 import phone from './assets/phone.png';
 import menu from './assets/menu.png';
+
 export const slides = [
   {
-    name: "Create Account",
+    name: 'Create Account',
     detail:
-      "Create/login to an existing account to get started. An account is created with your email and a desired password.",
-    img: phone
+      'Create/login to an existing account to get started. An account is created with your email and a desired password.',
+    img: phone,
   },
   {
-    name: "Log In",
+    name: 'Log In',
     detail:
-      "Create/login to an existing account to get started. An account is created with your email and a desired password.",
-    img: menu
+      'Create/login to an existing account to get started. An account is created with your email and a desired password.',
+    img: menu,
   },
   {
-    name: "Choose a Meal",
+    name: 'Choose a Meal',
     detail:
-      "Create/login to an existing account to get started. An account is created with your email and a desired password.",
-    img: phone
+      'Create/login to an existing account to get started. An account is created with your email and a desired password.',
+    img: phone,
   },
   {
-    name: "Make an Order",
+    name: 'Make an Order',
     detail:
-      "Create/login to an existing account to get started. An account is created with your email and a desired password.",
-    img: menu
+      'Create/login to an existing account to get started. An account is created with your email and a desired password.',
+    img: menu,
   },
 ];
